Avoid unneeded double precision math.

+++ src/refresh/gl/main.c
@@ -803,12 +803,12 @@ static void GL_InitTables(void)
         v = bytedirs[i];
         lat = acos(v[2]);
         lng = atan2(v[1], v[0]);
-        gl_static.latlngtab[i][0] = (int)(lat * (255.0f / (2 * M_PI))) & 255;
-        gl_static.latlngtab[i][1] = (int)(lng * (255.0f / (2 * M_PI))) & 255;
+        gl_static.latlngtab[i][0] = (int)(lat * (float)(255 / (2 * M_PI))) & 255;
+        gl_static.latlngtab[i][1] = (int)(lng * (float)(255 / (2 * M_PI))) & 255;
     }
 
     for (i = 0; i < 256; i++) {
-        gl_static.sintab[i] = sin(i * (2 * M_PI / 255.0f));
+        gl_static.sintab[i] = sin(i * (2 * M_PI / 255));
     }
 }
